Refactors the AppCenter uploading logic into its own class (out of AppCenterRest)
